pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gruener-podcast-feed"
version = "0.1.0"
description = "Newsletter-to-podcast pipeline for GruenGeschnackt"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "openai>=1.70.0,<2.0.0",
  "pydub>=0.25.1,<1.0.0",
  "python-dotenv>=1.1.0,<2.0.0",
]

[project.scripts]
gruenpod = "gruener_podcast_feed.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
